Remove season metadata.
At some point during early 2022, Ben announced that every 250 episodes of ILoF
would be considered one "season". On that basis I added season labels to all
the existing episodes, and wired it up to the summary pages.

As of Ep. 685, however, the crew have announced that we are now in the second
season of ILoF, following the cessation of regularly-schedule episodes.
Since the latter makes more sense anyway, I've removed all the previous season
metadata.